# backend/src/config/swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "crud-assign-cs",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/roles": {
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "Create Role",
        "description": "Create Role",
        "operationId": "createRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roleName": {
                    "type": "string",
                    "example": "Manager"
                  }
                }
              },
              "examples": {
                "Create Role": {
                  "value": {
                    "roleName": "Manager"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create User",
        "description": "Create User",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "admins@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456"
                  },
                  "roleId": {
                    "type": "string",
                    "example": "b274e3c1-9813-464a-8a1a-406315409692"
                  },
                  "username": {
                    "type": "string",
                    "example": "adminsupport"
                  }
                }
              },
              "examples": {
                "Create User": {
                  "value": {
                    "email": "admins@gmail.com",
                    "password": "123456",
                    "roleId": "b274e3c1-9813-464a-8a1a-406315409692",
                    "username": "adminsupport"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/users/2dd38e88-4718-48bb-8bbe-09329089746c": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update User By Id",
        "description": "Update User By Id",
        "operationId": "updateUserById",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "caterine@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "123455 "
                  },
                  "username": {
                    "type": "string",
                    "example": "Caterine41"
                  }
                }
              },
              "examples": {
                "Update User By Id": {
                  "value": {
                    "email": "caterine@gmail.com",
                    "password": "123455 ",
                    "username": "Caterine41"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/address": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "Get All Address",
        "description": "Get All Address",
        "operationId": "getAllAddress",
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Addresses"
        ],
        "summary": "Create Address",
        "description": "Create Address",
        "operationId": "createAddress",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "city": {
                    "type": "string",
                    "example": "Conroe"
                  },
                  "country": {
                    "type": "string",
                    "example": "Indonesia"
                  },
                  "customerId": {
                    "type": "string",
                    "example": "2ab00272-72e0-4c44-8dee-85fb21f61fcd"
                  },
                  "postalCode": {
                    "type": "string",
                    "example": "57234"
                  },
                  "province": {
                    "type": "string",
                    "example": "DK Jakarta"
                  },
                  "street": {
                    "type": "string",
                    "example": "Edwin Manor"
                  }
                }
              },
              "examples": {
                "Create Address": {
                  "value": {
                    "city": "Conroe",
                    "country": "Indonesia",
                    "customerId": "2ab00272-72e0-4c44-8dee-85fb21f61fcd",
                    "postalCode": "57234",
                    "province": "DK Jakarta",
                    "street": "Edwin Manor"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/contacts/0206be96-89f4-4c4b-bdbb-12e62bcf5aba": {
      "put": {
        "tags": [
          "Contacts"
        ],
        "summary": "Update Contact By Id",
        "description": "Update Contact By Id",
        "operationId": "updateContactById",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contactType": {
                    "type": "string",
                    "example": "Instagram"
                  },
                  "contactValue": {
                    "type": "string",
                    "example": "Marian_Schulist20@instagram.com"
                  },
                  "customerId": {
                    "type": "string",
                    "example": "7e4ea91a-5295-4301-807f-dabb7820e67a"
                  }
                }
              },
              "examples": {
                "Update Contact By Id": {
                  "value": {
                    "contactType": "Instagram",
                    "contactValue": "Marian_Schulist20@instagram.com",
                    "customerId": "7e4ea91a-5295-4301-807f-dabb7820e67a"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "summary": "Delete Contact By Id",
        "description": "Delete Contact By Id",
        "operationId": "deleteContactById",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
...
}